+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added `beforeRestore` and `afterRestore` hooks for plugin system
 - Added `LiveGame.playSound` to play a sound immediately and return the SoundToken
 - Added `useVoiceState` hook to access the voice state
+- Added `FixedAspectRatioContainer` Helper Component to create a container with a fixed aspect ratio
 
 ### _Incompatible Changes_